refactor: 역할에 따른 코드 분리
