+++ index.html
@@ -15,6 +15,6 @@
     <div id="root">
 
     </div>
-    <script src="app.ts" type="module"></script>
+    <script src="./src/app.ts" type="module"></script>
 </body>
 </html>
